+++ modules/manager/controller_setup.go
@@ -383,7 +383,7 @@ func SetupControllers(mgr manager.Manager, c *Config, cpsMgr *multiinstance.Mana
 		{
 			Enabled: c.GatewayControllerEnabled,
 			Controller: &gatewayclass.Reconciler{
-				CacheSyncTimeout:              c.CacheSyncTimeout,
+				ControllerOptions:             ctrlOpts,
 				Client:                        mgr.GetClient(),
 				LoggingMode:                   c.LoggingMode,
 				GatewayAPIExperimentalEnabled: c.GatewayAPIExperimentalEnabled,
@@ -393,7 +393,7 @@ func SetupControllers(mgr manager.Manager, c *Config, cpsMgr *multiinstance.Mana
 		{
 			Enabled: c.GatewayControllerEnabled,
 			Controller: &gateway.Reconciler{
-				CacheSyncTimeout:        c.CacheSyncTimeout,
+				ControllerOptions:       controllerOptions(ctrlOpts, withMaxConcurrentReconciles(int(c.MaxConcurrentReconcilesGateway))),
 				Client:                  mgr.GetClient(),
 				Scheme:                  mgr.GetScheme(),
 				Namespace:               c.ControllerNamespace,
@@ -408,8 +408,7 @@ func SetupControllers(mgr manager.Manager, c *Config, cpsMgr *multiinstance.Mana
 		{
 			Enabled: c.GatewayControllerEnabled || c.ControlPlaneControllerEnabled,
 			Controller: &controlplane.Reconciler{
-				CacheSyncPeriod:          c.CacheSyncPeriod,
-				CacheSyncTimeout:         c.CacheSyncTimeout,
+				ControllerOptions:        controllerOptions(ctrlOpts, withMaxConcurrentReconciles(int(c.MaxConcurrentReconcilesControlPlane))),
 				AnonymousReportsEnabled:  c.AnonymousReports,
 				LoggingMode:              c.LoggingMode,
 				Client:                   mgr.GetClient(),
@@ -432,7 +431,7 @@ func SetupControllers(mgr manager.Manager, c *Config, cpsMgr *multiinstance.Mana
 		{
 			Enabled: (c.DataPlaneControllerEnabled || c.GatewayControllerEnabled) && !c.DataPlaneBlueGreenControllerEnabled,
 			Controller: &dataplane.Reconciler{
-				CacheSyncTimeout:         c.CacheSyncTimeout,
+				ControllerOptions:        controllerOptions(ctrlOpts, withMaxConcurrentReconciles(int(c.MaxConcurrentReconcilesDataPlane))),
 				Client:                   mgr.GetClient(),
 				ClusterCASecretName:      c.ClusterCASecretName,
 				ClusterCASecretNamespace: c.ClusterCASecretNamespace,
@@ -450,6 +449,7 @@ func SetupControllers(mgr manager.Manager, c *Config, cpsMgr *multiinstance.Mana
 		{
 			Enabled: c.DataPlaneBlueGreenControllerEnabled,
 			Controller: &dataplane.BlueGreenReconciler{
+				ControllerOptions:        controllerOptions(ctrlOpts, withMaxConcurrentReconciles(int(c.MaxConcurrentReconcilesDataPlane))),
 				CacheSyncTimeout:         c.CacheSyncTimeout,
 				Client:                   mgr.GetClient(),
 				ClusterCASecretName:      c.ClusterCASecretName,
@@ -457,7 +457,7 @@ func SetupControllers(mgr manager.Manager, c *Config, cpsMgr *multiinstance.Mana
 				ClusterCAKeyConfig:       clusterCAKeyConfig,
 				SecretLabelSelector:      c.SecretLabelSelector,
 				DataPlaneController: &dataplane.Reconciler{
-					CacheSyncTimeout:         c.CacheSyncTimeout,
+					ControllerOptions:        controllerOptions(ctrlOpts, withMaxConcurrentReconciles(int(c.MaxConcurrentReconcilesDataPlane))),
 					Client:                   mgr.GetClient(),
 					ClusterCASecretName:      c.ClusterCASecretName,
 					ClusterCASecretNamespace: c.ClusterCASecretNamespace,
@@ -508,16 +508,16 @@ func SetupControllers(mgr manager.Manager, c *Config, cpsMgr *multiinstance.Mana
 		{
 			Enabled: c.AIGatewayControllerEnabled,
 			Controller: &specialized.AIGatewayReconciler{
-				CacheSyncTimeout: c.CacheSyncTimeout,
-				Client:           mgr.GetClient(),
-				LoggingMode:      c.LoggingMode,
+				ControllerOptions: ctrlOpts,
+				Client:            mgr.GetClient(),
+				LoggingMode:       c.LoggingMode,
 			},
 		},
 		// KongPluginInstallation controller
 		{
 			Enabled: c.KongPluginInstallationControllerEnabled,
 			Controller: &kongplugininstallation.Reconciler{
-				CacheSyncTimeout:       c.CacheSyncTimeout,
+				ControllerOptions:      ctrlOpts,
 				Client:                 mgr.GetClient(),
 				Scheme:                 mgr.GetScheme(),
 				LoggingMode:            c.LoggingMode,
@@ -528,7 +528,7 @@ func SetupControllers(mgr manager.Manager, c *Config, cpsMgr *multiinstance.Mana
 		{
 			Enabled: c.ControlPlaneExtensionsControllerEnabled,
 			Controller: &controlplane_extensions.Reconciler{
-				CacheSyncTimeout:                c.CacheSyncTimeout,
+				ControllerOptions:               ctrlOpts,
 				Client:                          mgr.GetClient(),
 				LoggingMode:                     c.LoggingMode,
 				DataPlaneScraperManagerNotifier: scrapersMgr,
@@ -539,13 +539,15 @@ func SetupControllers(mgr manager.Manager, c *Config, cpsMgr *multiinstance.Mana
 	// Konnect controllers
 	if c.KonnectControllersEnabled {
 		sdkFactory := sdkops.NewSDKFactory()
+		ctrlOpts := controllerOptions(ctrlOpts, withMaxConcurrentReconciles(int(c.MaxConcurrentReconcilesKonnect)))
+
 		controllerFactory := konnectControllerFactory{
-			sdkFactory:              sdkFactory,
-			loggingMode:             c.LoggingMode,
-			client:                  mgr.GetClient(),
-			syncPeriod:              c.KonnectSyncPeriod,
-			maxConcurrentReconciles: c.KonnectMaxConcurrentReconciles,
-			metricRecorder:          metricRecorder,
+			sdkFactory:        sdkFactory,
+			loggingMode:       c.LoggingMode,
+			client:            mgr.GetClient(),
+			syncPeriod:        c.KonnectSyncPeriod,
+			controllerOptions: ctrlOpts,
+			metricRecorder:    metricRecorder,
 		}
 
 		// Add additional Konnect controllers
@@ -555,6 +557,7 @@ func SetupControllers(mgr manager.Manager, c *Config, cpsMgr *multiinstance.Mana
 			ControllerDef{
 				Enabled: c.KonnectControllersEnabled,
 				Controller: konnect.NewKonnectAPIAuthConfigurationReconciler(
+					ctrlOpts,
 					sdkFactory,
 					c.LoggingMode,
 					mgr.GetClient(),
@@ -564,6 +567,7 @@ func SetupControllers(mgr manager.Manager, c *Config, cpsMgr *multiinstance.Mana
 			ControllerDef{
 				Enabled: c.KonnectControllersEnabled,
 				Controller: konnect.NewKongPluginReconciler(
+					ctrlOpts,
 					c.LoggingMode,
 					mgr.GetClient(),
 				),
@@ -572,6 +576,7 @@ func SetupControllers(mgr manager.Manager, c *Config, cpsMgr *multiinstance.Mana
 			ControllerDef{
 				Enabled: c.KonnectControllersEnabled,
 				Controller: konnect.NewKongCredentialSecretReconciler(
+					ctrlOpts,
 					c.LoggingMode,
 					mgr.GetClient(),
 					mgr.GetScheme(),
@@ -581,6 +586,7 @@ func SetupControllers(mgr manager.Manager, c *Config, cpsMgr *multiinstance.Mana
 			ControllerDef{
 				Enabled: (c.DataPlaneControllerEnabled || c.DataPlaneBlueGreenControllerEnabled) && c.KonnectControllersEnabled,
 				Controller: &konnect.KonnectExtensionReconciler{
+					ControllerOptions:        ctrlOpts,
 					SdkFactory:               sdkFactory,
 					LoggingMode:              c.LoggingMode,
 					Client:                   mgr.GetClient(),
@@ -658,12 +664,12 @@ func SetupControllers(mgr manager.Manager, c *Config, cpsMgr *multiinstance.Mana
 }
 
 type konnectControllerFactory struct {
-	sdkFactory              sdkops.SDKFactory
-	loggingMode             logging.Mode
-	client                  client.Client
-	syncPeriod              time.Duration
-	maxConcurrentReconciles uint
-	metricRecorder          metrics.Recorder
+	sdkFactory        sdkops.SDKFactory
+	loggingMode       logging.Mode
+	client            client.Client
+	syncPeriod        time.Duration
+	metricRecorder    metrics.Recorder
+	controllerOptions controller.Options
 }
 
 func newKonnectEntityController[
@@ -677,8 +683,8 @@ func newKonnectEntityController[
 			f.loggingMode,
 			f.client,
 			konnect.WithKonnectEntitySyncPeriod[T, TEnt](f.syncPeriod),
-			konnect.WithKonnectMaxConcurrentReconciles[T, TEnt](f.maxConcurrentReconciles),
 			konnect.WithMetricRecorder[T, TEnt](f.metricRecorder),
+			konnect.WithControllerOptions[T, TEnt](f.controllerOptions),
 		),
 	}
 }
@@ -690,6 +696,7 @@ func newKonnectPluginController[
 	return ControllerDef{
 		Enabled: true,
 		Controller: konnect.NewKonnectEntityPluginReconciler[T, TEnt](
+			f.controllerOptions,
 			f.loggingMode,
 			f.client,
 		),
@@ -700,5 +707,19 @@ func newGatewayAPIHybridController[t converter.RootObject, tPtr converter.RootOb
 	return ControllerDef{
 		Enabled:    true,
 		Controller: hybridgateway.NewHybridGatewayReconciler[t, tPtr](mgr, referenceGrantEnabled, fqdnMode, clusterDomain),
+	}
+}
+
+func controllerOptions(base controller.Options, opts ...func(*controller.Options)) controller.Options {
+	ret := base
+	for _, o := range opts {
+		o(&ret)
+	}
+	return ret
+}
+
+func withMaxConcurrentReconciles(n int) func(*controller.Options) {
+	return func(o *controller.Options) {
+		o.MaxConcurrentReconciles = n
 	}
 }