+++ CHANGELOG.md
@@ -159,6 +159,15 @@
   The finalizer `konnect.konghq.com/konnectapiauth-in-use` is automatically added
   when references exist and removed when all referencing resources are deleted.
   [#2726](https://github.com/Kong/kong-operator/pull/2726)
+- Add the following configuration flags for setting the maximum number of concurrent
+  reconciliation requests that can be processed by each controller group:
+  - `--max-concurrent-reconciles-dataplane-controller` for DataPlane controllers.
+  - `--max-concurrent-reconciles-controlplane-controller` for ControlPlane controllers.
+  - `--max-concurrent-reconciles-gateway-controller` for Gateway controllers.
+
+  NOTE: Konnect entities controllers still respect the
+  `--konnect-controller-max-concurrent-reconciles` flag.
+  [#2652](https://github.com/Kong/kong-operator/pull/2652)
 
 ### Changed